Moved safety structure to workorder directory
