+++ src/main/java/ch/cern/eam/wshub/core/client/InforClient.java
@@ -18,8 +18,8 @@ import ch.cern.eam.wshub.core.services.grids.GridsService;
 import ch.cern.eam.wshub.core.services.grids.impl.GridsServiceImpl;
 import ch.cern.eam.wshub.core.services.material.*;
 import ch.cern.eam.wshub.core.services.material.impl.*;
-import ch.cern.eam.wshub.core.services.safety.SafetyService;
-import ch.cern.eam.wshub.core.services.safety.impl.SafetyServiceImpl;
+import ch.cern.eam.wshub.core.services.workorders.SafetyService;
+import ch.cern.eam.wshub.core.services.workorders.impl.SafetyServiceImpl;
 import ch.cern.eam.wshub.core.services.workorders.*;
 import ch.cern.eam.wshub.core.services.workorders.impl.*;
 import ch.cern.eam.wshub.core.services.userdefinedscreens.UserDefinedTableService;
